app/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
boto3==1.35.24
botocore==1.35.24
certifi==2024.2.2
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.0.0
fastapi==0.109.2
flake8==7.0.0
flake8-print==5.0.0
freezegun==1.5.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ijson==3.3.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycodestyle==2.11.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyflakes==3.2.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==3.3.0
python-multipart==0.0.8
redis==5.0.2
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
StrEnum==0.4.15
tomli==2.0.1
typing_extensions==4.13.0
urllib3==2.2.3
uvicorn==0.27.0.post1
WTForms==3.1.2
